% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_auc_mvpls.R
\name{cv_auc_mvpls}
\alias{cv_auc_mvpls}
\title{CV multivariate PLS (comparable to cv_unique_auc_par) using AUC.}
\usage{
cv_auc_mvpls(
  X,
  Y,
  ncomp = min(10, ncol(X)),
  center = TRUE,
  folds = 5,
  verbose = TRUE
)
}
\arguments{
\item{X}{a number-of-observations times nodes matrix.}

\item{Y}{a number-of-observations times reponses matrix (only 0s and 1s).}

\item{ncomp}{number of components, integer.}

\item{center}{center the data.}

\item{folds}{a user defined list of folds (as generated by
caret::createFolds()) or an integer indicating the number of folds.
Independent of fdaPDE::FPCA.FEM's validation.}

\item{verbose}{logical, indicating if messages should be printed.}
}
\value{
A list of crossvalidates erros (CVEs).
}
\description{
CV multivariate PLS (comparable to cv_unique_auc_par) using AUC.
}
\examples{
# 1D example:

# library(pls)
# library(fda)

# Octane number:
Y <- as.matrix(pls::gasoline$octane)
Y <- ifelse(Y > 87, 1, 0)

# Gasoline NIR spectra:
X <- as.matrix( as.data.frame( pls::gasoline$NIR ) )

cv_mv <- cv_auc_mvpls(X = X, Y = Y, ncomp = 5, folds = 5, center = TRUE,
                  verbose = FALSE)
}
