% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls_tps.R
\name{fpls_tps}
\alias{fpls_tps}
\title{Penalized functional PLS based on the (2D) TPS representation of the data.
Inspired by Aguilera et al. 2016.}
\usage{
fpls_tps(
  X,
  Y,
  nodes,
  ncomp = 3,
  center = TRUE,
  nbasis,
  penalty = 0,
  tol = .Machine$double.eps^0.5,
  verbose = TRUE,
  stripped = FALSE,
  R0 = NULL,
  P = NULL,
  ...
)
}
\arguments{
\item{X}{a number of observations times nodes matrix.}

\item{Y}{a number of observations times reponses matrix.}

\item{nodes}{a set of x, y nodes.}

\item{ncomp}{number of components, integer.}

\item{center}{logical, indicating if data should be to centered.}

\item{nbasis}{number of TPS basis to use.}

\item{penalty}{penalty value, numeric.}

\item{tol}{convergence tolerance.}

\item{verbose}{logical, indicating if messages should be printed.}

\item{stripped}{logical.  If \code{TRUE} the calculations are stripped as
much as possible for speed; this is meant for use with cross-validation or
simulations when only the coefficients are needed.  Defaults to
\code{FALSE}. Inspired by package \code{pls}.}

\item{R0}{(mass) matrix of inner products between basis functions.}

\item{P}{penalty matrix.}

\item{...}{further arguments.  Currently not used}
}
\value{
an fpls-tps model.
}
\description{
Penalized functional PLS based on the (2D) TPS representation of the data.
Inspired by Aguilera et al. 2016.
}
\examples{
#' #' # Generate data (50 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 50, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
nodes <- L[["mesh"]][["nodes"]]

res_fpls_tps <- fpls_tps(X, Y, ncomp = 3, nodes = nodes, nbasis = 5,
                         penalty = 0, verbose = FALSE, stripped = FALSE )
}
