% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2d_data.R
\name{generate_2d_data}
\alias{generate_2d_data}
\title{2D data generation}
\usage{
generate_2d_data(x, y, num_samples = 100, beta_num = 3, Rsq = 0.95)
}
\arguments{
\item{x}{a sequence of nodes in first dimension.}

\item{y}{a sequence of nodes in second dimension.}

\item{num_samples}{number of samples.}

\item{beta_num}{an integer 1:5 of possible betas.}

\item{Rsq}{coefficient of determinaiton between a clean Y and a noisy Y.}
}
\value{
a list with the simulated data.
}
\description{
2D data generation
}
\examples{
x <- seq(0, 1, length.out = 20)
y <- seq(0, 1, length.out = 20)
L <- generate_2d_data(x, y, 60, 3, 0.95)
X <- L[["X"]]
Y <- L[["Y"]]
FEM_basis <- L[["basisobj"]]
mesh <- L[["mesh"]]
coeff_function <- L[["coefficient_function"]]
image(matrix(coeff_function, 20, 20), main = "Coefficient function")
}
