% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_3d_data.R
\name{generate_3d_data}
\alias{generate_3d_data}
\title{3D data generation}
\usage{
generate_3d_data(
  nodes,
  tetrahedrons,
  num_samples = 100,
  beta_num = 3,
  Rsq = 0.95
)
}
\arguments{
\item{nodes}{3D nodes (expanded grid).}

\item{tetrahedrons}{tetrahedrons as in fdaPDE.}

\item{num_samples}{number of samples.}

\item{beta_num}{an integer 1:3 of possible betas.}

\item{Rsq}{coefficient of determinaiton between a clean Y and a noisy Y.}
}
\value{
a list with the simulated data.
}
\description{
3D data generation
}
\examples{

library(fdaPDE)
data(sphere3Ddata)

nodes=sphere3Ddata$nodes
tetrahedrons=sphere3Ddata$tetrahedrons

new3d_data <- generate_3d_data(nodes = nodes, tetrahedrons = tetrahedrons,
num_samples = 30, beta_num = 3, Rsq = 0.9)

# library(plotly)
#
# gen_beta <- as.numeric(new3d_data[["coefficient_function"]] )
#
# fig_coeff <- plot_ly(x = ~nodes[, 1], y = ~nodes[, 2], z = ~nodes[, 3],
#                      marker = list(color = ~ gen_beta,
#                      colorscale = c('#FFE1A1', '#683531'),
#                      showscale = TRUE))
# fig_coeff <- fig_coeff \%>\% add_markers( alpha = 0.2)
# fig_coeff
#
#
# X1 <- as.numeric(new3d_data[["X"]][1, ] ) # first observation
#
# fig_X1 <- plot_ly(x = ~nodes[, 1], y = ~nodes[, 2], z = ~nodes[, 3],
#                      marker = list(color = ~ X1, colorscale = c('#FFE1A1',
#                       '#683531'), showscale = TRUE))
# fig_X1 <- fig_X1 \%>\% add_markers( alpha = 0.2)
# fig_X1
}
