% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fpls_tps}
\alias{predict.fpls_tps}
\title{Predict new data for FPLS-TPS model (penalized FPLS algorithm using a TPS basis,
inspired by Aguilera et al. 2016 methodology)}
\usage{
\method{predict}{fpls_tps}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{fpls_tps} model.}

\item{newdata}{new data matrix.}

\item{...}{further arguments.  Currently not used}
}
\value{
predicted values
}
\description{
Predict new data for FPLS-TPS model (penalized FPLS algorithm using a TPS basis,
inspired by Aguilera et al. 2016 methodology)
}
\examples{
#' # Generate data (50 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 50, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
nodes <- L[["mesh"]][["nodes"]]

res_fpls_tps <- fpls_tps(X, Y, ncomp = 3, nodes = nodes, nbasis = 5,
                         penalty = 0, verbose = FALSE, stripped = FALSE )

Y_pred <- predict(res_fpls_tps, X)
Y_hat <- fitted.values(res_fpls_tps)
plot(Y_pred, Y_hat)
abline(0, 1)
}
