% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.r1fpls_bs}
\alias{predict.r1fpls_bs}
\title{Predict new data for FPLS-FDA model (R1FPLS algorithm using an fda package basis)}
\usage{
\method{predict}{r1fpls_bs}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{fpls_fda} model.}

\item{newdata}{new data matrix.}

\item{...}{further arguments.  Currently not used}
}
\value{
predicted values
}
\description{
Predict new data for FPLS-FDA model (R1FPLS algorithm using an fda package basis)
}
\examples{
# 1D example:
# library(pls)
# library(fda)

# Octane number:
Y <- as.matrix( pls::gasoline$octane )

# Gasoline NIR spectra:
X <- as.matrix( as.data.frame( pls::gasoline$NIR ) )

# Wavenumber:
argvals <- seq(900, 1700, by = 2)

# Ruppert's law:  nbasis = nbreaks + norder - 2  and norder = degree + 1
n_breaks <- min(round(length(argvals)/4), 40)
n_basis <- n_breaks + (3+1) - 2

# B-spline basis:
bs_basis <- fda::create.bspline.basis(rangeval = range(argvals),
                                 nbasis = n_basis)



res_ps <-  r1fpls_bs(X = X, Y = Y, argvals = argvals,
                ncomp = 3, center = TRUE, penalty = 100,
                basisobj = bs_basis, stripped = FALSE)

predict(res_ps, newdata = X)
}
