% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.r1fpls_fem}
\alias{predict.r1fpls_fem}
\title{Predict new data for FPLS-FEM model (R1FPLS algorithm using FE basis)}
\usage{
\method{predict}{r1fpls_fem}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{fpls_fem} model.}

\item{newdata}{new data matrix.}

\item{...}{further arguments.  Currently not used}
}
\value{
predicted values
}
\description{
Predict new data for FPLS-FEM model (R1FPLS algorithm using FE basis)
}
\examples{
#' # Generate data (50 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 50, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
FEM_basis <- L[["basisobj"]]

res_fpls_fem <- r1fpls_fem(X, Y, ncomp = 3, center = TRUE,
                             basisobj = FEM_basis, penalty = 10,
                             verbose = FALSE, stripped = FALSE )

Y_pred <- predict(res_fpls_fem, X)
Y_hat <- fitted.values(res_fpls_fem)
}
