% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1fpls_fem.R
\name{r1fpls_fem}
\alias{r1fpls_fem}
\title{Penalized rank-one functional PLS using FE basis functions}
\usage{
r1fpls_fem(
  X,
  Y,
  ncomp = 3,
  center = TRUE,
  basisobj,
  penalty = 0,
  tol = .Machine$double.eps^0.5,
  verbose = TRUE,
  stripped = FALSE,
  ...
)
}
\arguments{
\item{X}{a number of observations times nodes matrix.}

\item{Y}{a number of observations times reponses matrix.}

\item{ncomp}{number of components, integer.}

\item{center}{logical, indicating if data should be to centered.}

\item{basisobj}{a Finite Elements basis as in the fdaPDE package.}

\item{penalty}{a vector of size ncomp indicating the penalty for each component.}

\item{tol}{convergence tolerance.}

\item{verbose}{logical, indicating if messages should be printed.}

\item{stripped}{logical.  If \code{TRUE} the calculations are stripped as
much as possible for speed; this is meant for use with cross-validation or
simulations when only the coefficients are needed.  Defaults to
\code{FALSE}. Inspired by package \code{pls}.}

\item{...}{further arguments.  Currently not used}
}
\value{
an r1fpls_fem model.
}
\description{
Penalized rank-one functional PLS using FE basis functions
}
\examples{
# Generate data (50 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 50, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
FEM_basis <- L[["basisobj"]]

results_fpls_fem <- r1fpls_fem(X, Y, ncomp = 3, center = TRUE,
                             basisobj = FEM_basis, penalty = 10,
                             verbose = TRUE )

true_coeff <- L[["coefficient_function"]]
image(matrix(true_coeff, 10, 10), main = "True coefficient function")

image(matrix(results_fpls_fem[["coefficient_function"]][ , , 3], 10, 10),
      main = "Estimated coefficient function")
}
